LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/default-src/manifest-v2-default-src-insecure-scheme.html

<!DOCTYPE html><!-- webkit-test-runner [ contentSecurityPolicyExtensionMode=v2 ] -->
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="default-src 'self' http://localhost:8000 'nonce-test';">
    <p><br>Test that CSP for extensions blocks sources with an insecure scheme.</p>
</head>
<body>
    <div id="result" text="FAIL">
      PASS
    </div>
    <script nonce='test'>
        if (window.testRunner) {
            testRunner.waitUntilDone();
            testRunner.dumpAsText();
        }

        let script = document.createElement('script');
        script.src = "http://localhost:8000/security/contentSecurityPolicy/resources/script.js";

        script.onload = function() {
            alert("FAIL: Should not have loaded script with an insecure scheme.");
            testRunner.notifyDone();
        };

        script.onerror = function() {
            alert("PASS: Successfully blocked script.");
            testRunner.notifyDone();
        };

        document.body.append(script);
    </script>
</body>
</html>
